// src/app/auth/schemas/schemas.ts
import * as z from "zod";

export const loginSchema = z.object({
    email: z.string().min(1, "Email is required").email("Enter a valid email"),
    password: z.string().min(8, "Password must be at least 8 characters"),
    remember: z.boolean().optional(),
})

export type LoginFormValues = z.infer<typeof loginSchema>